Add feature folder and reorganize food
