+++ src/features/common/Container.tsx
@@ -1,0 +1,9 @@
+import { type PropsWithChildren } from "react";
+
+export default function Container({ children }: PropsWithChildren) {
+  return (
+    <div className="m-auto box-border flex h-full w-full flex-col p-2 sm:max-w-2xl">
+      {children}
+    </div>
+  );
+}
