+++ src/features/food/components/Food/components/Consume.tsx
@@ -4,7 +4,7 @@ import { ArrowBigDownDash } from "lucide-react";
 import { api } from "~/utils/api";
 import type { ActionProps } from "../types";
 import { ACTIONS } from "../constants";
-import Modal from "~/components/common/Modal";
+import Modal from "~/features/common/Modal";
 
 export const Consume: FC<ActionProps> = ({
   data: { ammount, id, ubicationId },
